Added pagination support with local caching using sqldelight & store in playlist-detail, Added Playlist Detail UI

+++ shared/feature-playlist-detail/src/commonMain/kotlin/com/spotify/app/feature_playlist_detail/shared/data/dto/PlaylistDetailDTO.kt
@@ -9,11 +9,11 @@ data class PlaylistDetailDTO(
     val items: List<PlaylistDetailItemDTO>,
 
     @SerialName("limit")
-    val limit: Int,
+    val limit: Long,
 
     @SerialName("offset")
-    val offset: Int,
+    val offset: Long,
 
     @SerialName("total")
-    val total: Int
+    val total: Long
 )
