+++ shared/feature-playlist-detail/src/commonMain/kotlin/com/spotify/app/feature_playlist_detail/shared/domain/model/FetchPlaylistDetailRequest.kt
@@ -2,6 +2,6 @@ package com.spotify.app.feature_playlist_detail.shared.domain.model
 
 data class FetchPlaylistDetailRequest(
     val playlistId: String,
-    val limit: Int,
-    val offset: Int
+    val limit: Long,
+    val offset: Long
 )
