+++ shared/feature-playlist-detail/src/commonMain/kotlin/com/spotify/app/feature_playlist_detail/shared/domain/model/PlaylistDetail.kt
@@ -2,7 +2,7 @@ package com.spotify.app.feature_playlist_detail.shared.domain.model
 
 data class PlaylistDetail(
     val items: List<PlaylistDetailItem>,
-    val limit: Int,
-    val offset: Int,
-    val total: Int
+    val limit: Long,
+    val offset: Long,
+    val total: Long
 )
